requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
outcome==1.2.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
pyparsing==3.1.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
uritemplate==4.1.1
urllib3==1.26.16
utils==1.0.1
wsproto==1.2.0
